Wrap FunctionInvokingChatClient with CachingChatClient instead of the other way around (#132)
Response caching works reliably if a `FunctionInvokingChatClient` is wrapped with a `CachingChatClient`. The entire response (including cached function calls and results) are fetched from the cache in this case.

However, if a `CachingChatClient` is wrapped with a `FunctionInvokingChatClient` then caching only works partially. The initial tool call response message is fulfilled from the cache - but subsequent messages result in cache misses. This is because the function `callId`s are not stable between independent runs even when the conversations are identical.

Now, it is debatable whether function calls and results should be cached at all (since tool implementations can change between independent runs and fetching responses from a cache could end up hiding / delaying discovery of product truth). However, in the specific case of the current evaluation examples, we want all unit tests to execute quickly on retry so that we can demonstrate that the caching works.